requirements.txt
Django==2.2.4
djangorestframework==3.10.2
drf-writable-nested==0.6.1
pycountry==19.8.18